Move SigningKey
