+++ libkbfs/fake_key.go
@@ -22,9 +22,27 @@ func makeFakeRandomBytes(seed string, byteCount int) []byte {
 	return []byte(seed[:byteCount])
 }
 
+// SigningKeySecretSize is the size of a SigningKeySecret.
+const SigningKeySecretSize = libkb.NaclSigningKeySecretSize
+
+// SigningKeySecret is a secret that can be used to construct a SigningKey.
+type SigningKeySecret struct {
+	secret [SigningKeySecretSize]byte
+}
+
+// makeSigningKey makes a new Nacl signing key from the given secret.
+func makeSigningKey(secret SigningKeySecret) (kbfscrypto.SigningKey, error) {
+	kp, err := libkb.MakeNaclSigningKeyPairFromSecret(secret.secret)
+	if err != nil {
+		return kbfscrypto.SigningKey{}, err
+	}
+
+	return kbfscrypto.NewSigningKey(kp), nil
+}
+
 // MakeFakeSigningKeyOrBust makes a new signing key from fake
 // randomness made from the given seed.
-func MakeFakeSigningKeyOrBust(seed string) SigningKey {
+func MakeFakeSigningKeyOrBust(seed string) kbfscrypto.SigningKey {
 	fakeRandomBytes := makeFakeRandomBytes(seed, SigningKeySecretSize)
 	var fakeSecret SigningKeySecret
 	copy(fakeSecret.secret[:], fakeRandomBytes)
